+++ src/main/java/fr/estecka/variantscit/itemdata/transforms/IDataConversions.java
@@ -1,13 +1,15 @@
 package fr.estecka.variantscit.itemdata.transforms;
 
 import java.util.List;
+import java.util.Map;
 import java.util.function.Function;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import com.mojang.serialization.Codec;
 import fr.estecka.variantscit.CodecUtil;
-import fr.estecka.variantscit.VCitRegistries;
 import fr.estecka.variantscit.itemdata.containers.IDataContainer;
 import fr.estecka.variantscit.itemdata.containers.RawDataContainer;
+import fr.estecka.variantscit.itemdata.transforms.impl.AlternativeTransform;
 import net.minecraft.nbt.DoubleTag;
 import net.minecraft.nbt.NbtOps;
 import net.minecraft.nbt.NumericTag;
@@ -29,40 +31,23 @@ extends IDataTransform
 	}
 
 
+	static public final Codec<IDataTransform> EXPECT_UNIT_CODEC = CodecUtil.Enum(Codec.STRING, Map.of(
+		"string",          IDataConversions::StrictString,
+		"identifier",      IDataConversions::StricIdentifier,
+		"number",          IDataConversions::StrictNumber,
+		"rich_text",       IDataConversions::StrictRichText,
+		"rich_text_array", IDataConversions::StrictRichTextArray
+	));
 
-	@Deprecated static public final Codec<IDataConversions<?>> LEGACY_CODEC = VCitRegistries.NBT_INPUTS.codec;
-	@Deprecated static public final Codec<IDataConversions<?>[]> LEGACY_ARRAY_CODEC = CodecUtil.OneOrMany(CODEC).xmap(list->list.toArray(IDataConversions[]::new), array->List.of(array));
-	@Deprecated static public final Codec<IDataConversions<?>> LEGACY_GROUP_CODEC = LEGACY_ARRAY_CODEC.xmap(IDataConversions::Grouped, type -> type instanceof Group group ? group.content() : new IDataConversions[]{ type } );
+	static public final Codec<IDataTransform> EXPECT_GROUP_CODEC = CodecUtil.OneOrMany(EXPECT_UNIT_CODEC)
+		.xmap(AlternativeTransform::Wrap, AlternativeTransform::Unwrap)
+		;
 
-	static public IDataConversions<String> TO_STRING = Alternatives(
+	static public IDataConversions<String> AGGRESSIVE_TO_STRING = Alternatives(
 		IDataConversions::SoftCastToString,
 		o->SoftCastToString(SoftCastToText(o)),
 		o->SoftCastToString(LooseCastTextArray(o))
 	);
-
-	@Deprecated
-	public record Group(IDataConversions<?>... content)
-	implements IDataConversions<Object>
-	{
-		@Override
-		public Object flatLooseTransform(Object value) {
-			for (int i=0; i<content.length; ++i){
-				Object result = content[i].flatLooseTransform(value);
-				if (result != null) return result;
-			}
-			return null;
-		}
-
-	}
-
-	@Deprecated
-	static public IDataConversions<?> Grouped(IDataConversions<?>... group){
-		if (group.length == 1)
-			return group[0];
-		else
-			return new Group(group);
-	}
-
 
 	@SafeVarargs
 	static public <T> IDataConversions<T> Alternatives(Function<Object,T>... functions){
@@ -74,6 +59,62 @@ extends IDataTransform
 			return null;
 		};
 	}
+
+
+/******************************************************************************/
+/* # "Expect" field                                                           */
+/******************************************************************************/
+
+static public IDataContainer StricIdentifier(IDataContainer input) {
+	if (input.value() instanceof ResourceLocation id)
+		return input;
+	if (input.value() instanceof String string)
+		return RawDataContainer.OfNullable(ResourceLocation.tryParse(string));
+	if (input.asNbt() instanceof StringTag nbt)
+		return RawDataContainer.OfNullable(ResourceLocation.tryParse(nbt.getAsString()));
+	return null;
+}
+
+static public IDataContainer StrictString(IDataContainer input) {
+	if (input.value() instanceof String)
+		return input;
+	if (input.value() instanceof ResourceLocation id)
+		return RawDataContainer.OfNullable(input.asString());
+	if (input.asNbt() instanceof StringTag nbt)
+		return RawDataContainer.OfNullable(nbt.getAsString());
+	return null;
+}
+
+static public IDataContainer StrictNumber(IDataContainer input) {
+	if (input.value() instanceof Number)
+		return input;
+	if (input.asNbt() instanceof NumericTag nbt)
+		return RawDataContainer.OfNullable(nbt.getAsNumber());
+
+	return null;
+}
+
+static public IDataContainer StrictRichText(IDataContainer input) {
+	if (input.value() instanceof Component)
+		return input;
+	
+	var nbt = input.asNbt();
+	if (nbt != null)
+		return RawDataContainer.OfNullable(NbtToText(nbt));
+	else
+		return null;
+}
+
+static public IDataContainer StrictRichTextArray(IDataContainer input) {
+	if (input.value() instanceof Component)
+		return input;
+	
+	var nbt = input.asNbt();
+	if (nbt != null)
+		return RawDataContainer.OfNullable(NbtToText(nbt));
+	else
+		return null;
+}
 
 
 /******************************************************************************/
@@ -124,14 +165,14 @@ extends IDataTransform
 /******************************************************************************/
 /* # Rich Text Util                                                           */
 /******************************************************************************/
-	static public Component NbtToText(Tag nbt){
+	static public Component NbtToText(@NotNull Tag nbt){
 		return ComponentSerialization.FLAT_CODEC
 			.parse(NbtOps.INSTANCE, nbt)
 			.mapOrElse(Function.identity(), _0->null)
 			;
 	}
 
-	static public List<Component> NbtToTextArray(Tag nbt){
+	static public List<Component> NbtToTextArray(@NotNull Tag nbt){
 		return ComponentSerialization.FLAT_CODEC
 			.sizeLimitedListOf(256)
 			.parse(NbtOps.INSTANCE, nbt)
@@ -139,7 +180,7 @@ extends IDataTransform
 			;
 	}
 
-	static public String TextArrayToString(List<Component> lines){
+	static public String TextArrayToString(@NotNull List<Component> lines){
 		StringBuilder builder = new StringBuilder();
 		for (var l : lines) {
 			builder.append(l.getString());